+++ packages/container/libraries/core/src/metadata/models/InjectFromHierarchyOptions.ts
@@ -1,0 +1,4 @@
+export interface InjectFromHierarchyOptions {
+  extendConstructorArguments?: boolean | undefined;
+  extendProperties?: boolean | undefined;
+}